Merge, particularly including connectivity extraction updates
