+++ layout21raw/src/lib.rs
@@ -1,25 +1,31 @@
 //!
 //! # Raw Layout
 //!
-//! The most general, most flexible representation layer in layout21.
+//! The most general and lowest-level representation layer in layout21.
 //! Consists of geometric primitives and instances of other layout cells,
 //! much akin to nearly any legacy layout system.
 //!
 
-// Std-Lib
-use std::collections::{HashMap, HashSet};
-use std::convert::{TryFrom, TryInto};
-use std::hash::Hash;
-use std::ops::Not;
+// Crates.io dependencies, at crate-level for their macros 
+#[macro_use]
+extern crate enum_dispatch;
 
-// Crates.io
-use serde::{Deserialize, Serialize};
-use slotmap::{new_key_type, SlotMap};
+// Internal modules
+pub mod bbox;
+pub mod data;
+pub mod error;
+pub mod geom;
 
-// Internal modules & re-exports
+// Re-exports
+#[doc(inline)]
+pub use bbox::*;
+#[doc(inline)]
+pub use data::*;
+#[doc(inline)]
+pub use error::*;
+#[doc(inline)]
+pub use geom::*;
 pub use layout21utils as utils;
-use utils::ErrorContext;
-use utils::{Ptr, PtrList};
 
 // Optional-feature modules
 #[cfg(feature = "gds")]
@@ -28,800 +34,7 @@ pub mod gds;
 pub mod lef;
 #[cfg(feature = "proto")]
 pub mod proto;
+
 // Unit tests
 #[cfg(test)]
 mod tests;
-
-// Create key-types for each internal type stored in [SlotMap]s
-new_key_type! {
-    /// Keys for [Layer] entries
-    pub struct LayerKey;
-}
-/// LayoutError-Specific Result Type
-pub type LayoutResult<T> = Result<T, LayoutError>;
-
-///
-/// # Layout Error Enumeration
-///
-pub enum LayoutError {
-    /// Error Exporting to Foreign Format
-    Export {
-        message: String,
-        stack: Vec<ErrorContext>,
-    },
-    /// Error Importing from Foreign Format
-    Import {
-        message: String,
-        stack: Vec<ErrorContext>,
-    },
-    /// Conversion Errors, with Boxed External Error
-    Conversion {
-        message: String,
-        err: Box<dyn std::error::Error>,
-        stack: Vec<ErrorContext>,
-    },
-    /// Boxed External Errors
-    Boxed(Box<dyn std::error::Error>),
-    /// Uncategorized Error, with String Message
-    Str(String),
-    /// # [Ptr] Locking
-    /// Caused by trouble with a [Ptr]: either deadlock, or panic while holding a lock.
-    /// Generally caused by a [std::sync::PoisonError], which is not forwardable due to lifetime constraints.
-    PtrLock,
-}
-impl LayoutError {
-    /// Create a [LayoutError::Message] from anything String-convertible
-    pub fn msg(s: impl Into<String>) -> Self {
-        Self::Str(s.into())
-    }
-    /// Create an error-variant [Result] of our [LayoutError::Message] variant from anything String-convertible
-    pub fn fail<T>(s: impl Into<String>) -> Result<T, Self> {
-        Err(Self::msg(s))
-    }
-}
-impl std::fmt::Debug for LayoutError {
-    /// Display a [LayoutError]
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            LayoutError::Export { message, stack } => {
-                write!(f, "Export Error: \n - {} \n - {:?}", message, stack)
-            }
-            LayoutError::Import { message, stack } => {
-                write!(f, "Import Error: \n - {} \n - {:?}", message, stack)
-            }
-            LayoutError::Conversion {
-                message,
-                err,
-                stack,
-            } => write!(
-                f,
-                "Conversion Error: \n - {} \n - {} \n - {:?}",
-                message, err, stack
-            ),
-            LayoutError::Boxed(err) => err.fmt(f),
-            LayoutError::Str(err) => err.fmt(f),
-            LayoutError::PtrLock => write!(f, "[std::sync::PoisonError]"),
-        }
-    }
-}
-impl std::fmt::Display for LayoutError {
-    /// Display a [LayoutError]
-    /// Delegates to the [Debug] implementation
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        std::fmt::Debug::fmt(self, f)
-    }
-}
-impl std::error::Error for LayoutError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        match self {
-            Self::Boxed(e) => Some(&**e),
-            _ => None,
-        }
-    }
-}
-
-impl From<String> for LayoutError {
-    fn from(s: String) -> Self {
-        Self::Str(s)
-    }
-}
-impl From<&str> for LayoutError {
-    fn from(s: &str) -> Self {
-        Self::Str(s.to_string())
-    }
-}
-impl From<std::num::TryFromIntError> for LayoutError {
-    fn from(e: std::num::TryFromIntError) -> Self {
-        Self::Boxed(Box::new(e))
-    }
-}
-impl From<utils::ser::Error> for LayoutError {
-    fn from(e: utils::ser::Error) -> Self {
-        Self::Boxed(Box::new(e))
-    }
-}
-impl<T> From<std::sync::PoisonError<T>> for LayoutError {
-    fn from(_e: std::sync::PoisonError<T>) -> Self {
-        Self::PtrLock
-    }
-}
-impl<T: std::error::Error + 'static> From<Box<T>> for LayoutError {
-    fn from(e: Box<T>) -> Self {
-        Self::Boxed(e)
-    }
-}
-/// Distance Units Enumeration
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
-pub enum Units {
-    /// Micrometers, or microns for we olde folke
-    Micro,
-    /// Nanometers
-    Nano,
-    /// Angstroms
-    Angstrom,
-    /// Picometers
-    Pico,
-}
-impl Default for Units {
-    /// Default units are nanometers
-    fn default() -> Units {
-        Units::Nano
-    }
-}
-/// Enumerated SI Units
-#[allow(dead_code)] // FIXME!
-enum SiUnits {
-    Yocto, // E-24
-    Zepto, // E-21
-    Atto,  // E-18
-    Femto, // E-15
-    Pico,  // E-12
-    Nano,  // E-9
-    Micro, // E-6
-    Milli, // E-3
-    Centi, // E-2
-    Deci,  // E-1
-    Deca,  // E1
-    Hecto, // E2
-    Kilo,  // E3
-    Mega,  // E6
-    Giga,  // E9
-    Tera,  // E12
-    Peta,  // E15
-    Exa,   // E18
-    Zetta, // E21
-    Yotta, // E24
-}
-impl Default for SiUnits {
-    /// Default units are nano-scale
-    #[allow(dead_code)] // FIXME!
-    fn default() -> SiUnits {
-        SiUnits::Nano
-    }
-}
-impl SiUnits {
-    /// Get the exponent of the unit
-    #[allow(dead_code)] // FIXME!
-    fn exp(&self) -> isize {
-        use SiUnits::*;
-        match self {
-            Yocto => -24,
-            Zepto => -21,
-            Atto => -18,
-            Femto => -15,
-            Pico => -12,
-            Nano => -9,
-            Micro => -6,
-            Milli => -3,
-            Centi => -2,
-            Deci => -1,
-            Deca => 1,
-            Hecto => 2,
-            Kilo => 3,
-            Mega => 6,
-            Giga => 9,
-            Tera => 12,
-            Peta => 15,
-            Exa => 18,
-            Zetta => 21,
-            Yotta => 24,
-        }
-    }
-}
-
-/// # Point in two-dimensional layout-space
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub struct Point {
-    pub x: isize,
-    pub y: isize,
-}
-impl Point {
-    /// Create a new [Point] from (x,y) coordinates
-    pub fn new(x: isize, y: isize) -> Self {
-        Self { x, y }
-    }
-    /// Create a new [Point] which serves as an offset in direction `dir`
-    pub fn offset(val: isize, dir: Dir) -> Self {
-        match dir {
-            Dir::Horiz => Self { x: val, y: 0 },
-            Dir::Vert => Self { x: 0, y: val },
-        }
-    }
-    /// Create a new point shifted by `x` in the x-dimension and by `y` in the y-dimension
-    pub fn shift(&self, p: &Point) -> Point {
-        Point {
-            x: p.x + self.x,
-            y: p.y + self.y,
-        }
-    }
-    /// Create a new point scaled by `p.x` in the x-dimension and by `p.y` in the y-dimension
-    pub fn scale(&self, p: &Point) -> Point {
-        Point {
-            x: p.x * self.x,
-            y: p.y * self.y,
-        }
-    }
-    /// Get the coordinate associated with direction `dir`
-    pub fn coord(&self, dir: Dir) -> isize {
-        match dir {
-            Dir::Horiz => self.x,
-            Dir::Vert => self.y,
-        }
-    }
-}
-/// Direction Enumeration
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
-pub enum Dir {
-    Horiz,
-    Vert,
-}
-impl Dir {
-    /// Whichever direction we are, return the other one.
-    pub fn other(self) -> Self {
-        match self {
-            Self::Horiz => Self::Vert,
-            Self::Vert => Self::Horiz,
-        }
-    }
-}
-impl Not for Dir {
-    type Output = Self;
-    /// Exclamation Operator returns the opposite direction
-    fn not(self) -> Self::Output {
-        Self::other(self)
-    }
-}
-
-/// Instance of another Cell
-#[derive(Debug, Clone)]
-pub struct Instance {
-    /// Instance Name
-    pub inst_name: String,
-    /// Cell Definition Reference
-    pub cell: Ptr<Cell>,
-    /// Location of `cell` origin
-    /// regardless of rotation or reflection
-    pub loc: Point,
-    /// Vertical reflection,
-    /// applied *before* rotation
-    pub reflect_vert: bool,
-    /// Angle of rotation (degrees),
-    /// Clockwise and applied *after* reflection
-    pub angle: Option<f64>,
-}
-
-/// Layer Set & Manager
-/// Keep track of active layers, and index them by name and number.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct Layers {
-    slots: SlotMap<LayerKey, Layer>,
-    nums: HashMap<i16, LayerKey>,
-    names: HashMap<String, LayerKey>,
-}
-impl Layers {
-    /// Add a [Layer] to our slot-map and number-map, and name-map
-    pub fn add(&mut self, layer: Layer) -> LayerKey {
-        // FIXME: conflicting numbers and/or names, at least some of which tend to happen, over-write each other.
-        // Sort out the desired behavior here.
-        let num = layer.layernum;
-        let name = layer.name.clone();
-        let key = self.slots.insert(layer);
-        self.nums.insert(num, key.clone());
-        if let Some(s) = name {
-            self.names.insert(s, key.clone());
-        }
-        key
-    }
-    /// Get the next-available (lowest) layer number
-    pub fn nextnum(&self) -> LayoutResult<i16> {
-        for k in 0..i16::MAX {
-            if !self.nums.contains_key(&k) {
-                return Ok(k);
-            }
-        }
-        LayoutError::fail("No more layer numbers available")
-    }
-    /// Get a reference to the [LayerKey] for layer-number `num`
-    pub fn keynum(&self, num: i16) -> Option<LayerKey> {
-        self.nums.get(&num).map(|x| x.clone())
-    }
-    /// Get a reference to the [LayerKey] layer-name `name`
-    pub fn keyname(&self, name: impl Into<String>) -> Option<LayerKey> {
-        self.names.get(&name.into()).map(|x| x.clone())
-    }
-    /// Get a reference to [Layer] number `num`
-    pub fn num(&self, num: i16) -> Option<&Layer> {
-        let key = self.nums.get(&num)?;
-        self.slots.get(*key)
-    }
-    /// Get a reference to [Layer] name `name`
-    pub fn name(&self, name: &str) -> Option<&Layer> {
-        let key = self.names.get(name)?;
-        self.slots.get(*key)
-    }
-    /// Get the name of `layerkey`
-    pub fn get_name(&self, layerkey: LayerKey) -> Option<&String> {
-        let layer = self.slots.get(layerkey)?;
-        layer.name.as_ref()
-    }
-    /// Get a reference to [Layer] from [LayerKey] `key`
-    pub fn get(&self, key: LayerKey) -> Option<&Layer> {
-        self.slots.get(key)
-    }
-    /// Get the ([LayerKey], [LayerPurpose]) objects for numbers (`layernum`, `purposenum`) if present.
-    /// Inserts a new [Layer] if `layernum` is not present.
-    /// Returns `LayerPurpose::Other(purposenum)` if `purposenum` is not present on that layer.
-    pub fn get_or_insert(
-        &mut self,
-        layernum: i16,
-        purposenum: i16,
-    ) -> LayoutResult<(LayerKey, LayerPurpose)> {
-        // Get the [LayerKey] for `layernum`, creating the [Layer] if it doesn't exist.
-        let key = match self.keynum(layernum) {
-            Some(key) => key.clone(),
-            None => self.add(Layer::from_num(layernum)),
-        };
-        // Slightly awkwardly, get that [Layer] (back), so we can get or add a [LayerPurpose]
-        let layer = self
-            .slots
-            .get_mut(key)
-            .ok_or(LayoutError::msg("Layer Not Found"))?;
-        // Get or create the corresponding [LayerPurpose]
-        let purpose = match layer.purpose(purposenum) {
-            Some(purpose) => purpose.clone(),
-            None => {
-                // Create a new anonymous/ numbered layer-purpose
-                let purpose = LayerPurpose::Other(purposenum);
-                layer.add_purpose(purposenum, purpose.clone())?;
-                purpose
-            }
-        };
-        Ok((key, purpose))
-    }
-    /// Get a shared reference to the internal <[LayerKey], [Layer]> map
-    pub fn slots(&self) -> &SlotMap<LayerKey, Layer> {
-        &self.slots
-    }
-}
-/// Layer-Purpose Enumeration
-/// Includes the common use-cases for each shape,
-/// and two "escape hatches", one named and one not.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
-pub enum LayerPurpose {
-    // First-class enumerated purposes
-    Drawing,
-    Pin,
-    Label,
-    Obstruction,
-    Outline,
-    /// Named purpose, not first-class supported
-    Named(String, i16),
-    /// Other purpose, not first-class supported nor named
-    Other(i16),
-}
-/// # Layer Specification
-/// As in seemingly every layout system, this uses two numbers to identify each layer.
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-pub struct LayerSpec(i16, i16);
-impl LayerSpec {
-    pub fn new(n1: i16, n2: i16) -> Self {
-        Self(n1, n2)
-    }
-}
-/// # Per-Layer Datatype Specification
-/// Includes the datatypes used for each category of element on layer `layernum`
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub struct Layer {
-    /// Layer Number
-    pub layernum: i16,
-    /// Layer Name
-    pub name: Option<String>,
-    /// Number => Purpose Lookup
-    purps: HashMap<i16, LayerPurpose>,
-    /// Purpose => Number Lookup
-    nums: HashMap<LayerPurpose, i16>,
-}
-impl Layer {
-    /// Create a new [Layer] with the given `layernum` and `name`
-    pub fn new(layernum: i16, name: impl Into<String>) -> Self {
-        Self {
-            layernum,
-            name: Some(name.into()),
-            ..Default::default()
-        }
-    }
-    /// Create a new [Layer] with the given `layernum`.
-    pub fn from_num(layernum: i16) -> Self {
-        Self {
-            layernum,
-            ..Default::default()
-        }
-    }
-    /// Create a new [Layer] purpose-numbers `pairs`.
-    pub fn from_pairs(layernum: i16, pairs: &[(i16, LayerPurpose)]) -> LayoutResult<Self> {
-        let mut layer = Self::from_num(layernum);
-        for (num, purpose) in pairs {
-            layer.add_purpose(*num, purpose.clone())?;
-        }
-        Ok(layer)
-    }
-    /// Add purpose-numbers `pairs`. Consumes and returns `self` for chainability.
-    pub fn add_pairs(mut self, pairs: &[(i16, LayerPurpose)]) -> LayoutResult<Self> {
-        for (num, purpose) in pairs {
-            self.add_purpose(*num, purpose.clone())?;
-        }
-        Ok(self)
-    }
-    /// Add a new [LayerPurpose]
-    pub fn add_purpose(&mut self, num: i16, purp: LayerPurpose) -> LayoutResult<()> {
-        // If we get a numbered purpose, make sure its id matches `num`.
-        match purp {
-            LayerPurpose::Named(_, k) | LayerPurpose::Other(k) => {
-                if k != num {
-                    LayoutError::fail("Invalid LayerPurpose")?;
-                }
-            }
-            _ => (),
-        };
-        self.purps.insert(num, purp.clone());
-        self.nums.insert(purp, num);
-        Ok(())
-    }
-    /// Retrieve purpose-number `num`
-    pub fn purpose(&self, num: i16) -> Option<&LayerPurpose> {
-        self.purps.get(&num)
-    }
-    /// Retrieve the purpose-number for this layer and [Purpose] `purpose`
-    pub fn num(&self, purpose: &LayerPurpose) -> Option<i16> {
-        self.nums.get(purpose).copied()
-    }
-}
-
-/// Raw Abstract-Layout
-/// Contains geometric [Element]s generally representing pins and blockages
-/// Does not contain instances, arrays, or layout-implementation details
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Abstract {
-    /// Cell Name
-    pub name: String,
-    /// Outline
-    pub outline: Element,
-    /// Ports
-    pub ports: Vec<AbstractPort>,
-    /// Blockages
-    pub blockages: HashMap<LayerKey, Vec<Shape>>,
-}
-impl Abstract {
-    /// Create a new [Abstract] with the given `name`
-    pub fn new(name: impl Into<String>, outline: Element) -> Self {
-        let name = name.into();
-        Self {
-            name,
-            outline,
-            ports: Vec::new(),
-            blockages: HashMap::new(),
-        }
-    }
-}
-/// # Port Element for [Abstract]s
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct AbstractPort {
-    /// Net Name
-    pub net: String,
-    /// Shapes, with paired [Layer] keys
-    pub shapes: HashMap<LayerKey, Vec<Shape>>,
-}
-impl AbstractPort {
-    /// Create a new [AbstractPort] with the given `name`
-    pub fn new(net: impl Into<String>) -> Self {
-        let net = net.into();
-        Self {
-            net,
-            shapes: HashMap::new(),
-        }
-    }
-}
-
-/// # Raw Layout Library  
-/// A collection of cell-definitions and sub-library definitions
-#[derive(Debug, Clone, Default)]
-pub struct Library {
-    /// Library Name
-    pub name: String,
-    /// Distance Units
-    pub units: Units,
-    /// Layer Definitions
-    pub layers: Ptr<Layers>,
-    /// Cell Definitions
-    pub cells: PtrList<Cell>,
-}
-impl Library {
-    /// Create a new and empty Library
-    pub fn new(name: impl Into<String>, units: Units) -> Self {
-        Self {
-            name: name.into(),
-            units,
-            ..Default::default()
-        }
-    }
-    /// Convert to a GDSII Library
-    #[cfg(feature = "gds")]
-    pub fn to_gds(&self) -> LayoutResult<gds::gds21::GdsLibrary> {
-        gds::GdsExporter::export(&self)
-    }
-    /// Create from GDSII
-    #[cfg(feature = "gds")]
-    pub fn from_gds(
-        gdslib: &gds::gds21::GdsLibrary,
-        layers: Option<Ptr<Layers>>,
-    ) -> LayoutResult<Library> {
-        gds::GdsImporter::import(&gdslib, layers)
-    }
-    /// Convert to ProtoBuf
-    #[cfg(feature = "proto")]
-    pub fn to_proto(&self) -> LayoutResult<proto::proto::Library> {
-        proto::ProtoExporter::export(&self)
-    }
-    /// Create from ProtoBuf, or anything convertible into a Proto Library
-    #[cfg(feature = "proto")]
-    pub fn from_proto<T>(plib: T, layers: Option<Ptr<Layers>>) -> LayoutResult<Library>
-    where
-        // These trait bounds aren't pretty, but more or less say:
-        // * T is convertible into [proto::proto::Library]
-        // * Its conversion's error-type is convertible into [LayoutError]
-        // The "Into" form of the second condition would be something like:
-        // `<T as TryInto<proto::proto::Library>::Error>: Into<LayoutError>`
-        // but doesn't quite work, whereas "constraining" [LayoutError] does.
-        T: TryInto<proto::proto::Library>,
-        LayoutError: From<<T as TryInto<proto::proto::Library>>::Error>,
-    {
-        let plib = plib.try_into()?;
-        proto::ProtoImporter::import(&plib, layers)
-    }
-}
-
-/// # Dependency-Orderer
-#[derive(Debug)]
-pub struct DepOrder<'lib> {
-    // FIXME: move to utils shared version
-    lib: &'lib Library,
-    stack: Vec<Ptr<Cell>>,
-    seen: HashSet<Ptr<Cell>>,
-}
-impl<'lib> DepOrder<'lib> {
-    fn order(lib: &'lib Library) -> Vec<Ptr<Cell>> {
-        let mut myself = Self {
-            lib,
-            stack: Vec::new(),
-            seen: HashSet::new(),
-        };
-        for cell in myself.lib.cells.iter() {
-            myself.push(cell);
-        }
-        myself.stack
-    }
-    fn push(&mut self, ptr: &Ptr<Cell>) {
-        // If the Cell hasn't already been visited, depth-first search it
-        if !self.seen.contains(&ptr) {
-            // Read the cell-pointer
-            let cell = ptr.read().unwrap();
-            // If the cell has an implementation, visit its [Instance]s before inserting it
-            if let Some(layout) = &cell.layout {
-                for inst in &layout.insts {
-                    self.push(&inst.cell);
-                }
-            }
-            // And insert the cell (pointer) itself
-            self.seen.insert(Ptr::clone(ptr));
-            self.stack.push(Ptr::clone(ptr));
-        }
-    }
-}
-
-/// Collection of the Views describing a Cell
-#[derive(Debug, Default, Clone)]
-pub struct Cell {
-    // Cell Name
-    pub name: String,
-    // Layout Abstract
-    pub abs: Option<Abstract>,
-    // Layout Implementation
-    pub layout: Option<Layout>,
-}
-impl Cell {
-    /// Create a new and empty Cell named `name`
-    pub fn new(name: impl Into<String>) -> Self {
-        let name = name.into();
-        Self {
-            name,
-            ..Default::default()
-        }
-    }
-}
-impl From<Abstract> for Cell {
-    fn from(src: Abstract) -> Self {
-        Self {
-            name: src.name.clone(),
-            abs: Some(src),
-            ..Default::default()
-        }
-    }
-}
-impl From<Layout> for Cell {
-    fn from(src: Layout) -> Self {
-        Self {
-            name: src.name.clone(),
-            layout: Some(src),
-            ..Default::default()
-        }
-    }
-}
-
-/// Raw-Layout Implementation
-#[derive(Debug, Clone, Default)]
-pub struct Layout {
-    /// Cell Name
-    pub name: String,
-    /// Instances
-    pub insts: Vec<Instance>,
-    /// Primitive/ Geometric Elements
-    pub elems: Vec<Element>,
-    /// Text Annotations
-    pub annotations: Vec<TextElement>,
-}
-/// # Text Annotation
-///
-/// Note [layout21::raw::TextElement]s are "layer-less",
-/// i.e. they do not sit on different layers,
-/// and do not describe connectivity or generate pins.
-/// These are purely annotations in the sense of "design notes".
-///
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub struct TextElement {
-    /// String Value
-    pub string: String,
-    /// Location
-    pub loc: Point,
-}
-/// # Primitive Geometric Element
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct Element {
-    /// Net Name
-    pub net: Option<String>,
-    /// Layer (Reference)
-    pub layer: LayerKey,
-    /// Purpose
-    pub purpose: LayerPurpose,
-    /// Shape
-    pub inner: Shape,
-}
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub enum Shape {
-    Rect { p0: Point, p1: Point },
-    Poly { pts: Vec<Point> },
-    Path { width: usize, pts: Vec<Point> },
-}
-impl Shape {
-    /// Retrieve our "origin", or first [Point]
-    pub fn point0(&self) -> &Point {
-        match *self {
-            Shape::Rect { ref p0, p1: _ } => p0,
-            Shape::Poly { ref pts } => &pts[0],
-            Shape::Path { ref pts, .. } => &pts[0],
-        }
-    }
-    /// Calculate our center-point
-    pub fn center(&self) -> Point {
-        match *self {
-            Shape::Rect { ref p0, ref p1 } => Point::new((p0.x + p1.x) / 2, (p0.y + p1.y) / 2),
-            Shape::Path { ref pts, .. } => {
-                // Place on the center of the first segment
-                let p0 = &pts[0];
-                let p1 = &pts[1];
-                Point::new((p0.x + p1.x) / 2, (p0.y + p1.y) / 2)
-            }
-            Shape::Poly { .. } => {
-                unimplemented!("Shape::Poly/Path::center");
-            }
-        }
-    }
-    /// Indicate whether this shape is (more or less) horizontal or vertical
-    /// Primarily used for orienting label-text
-    pub fn orientation(&self) -> Dir {
-        match *self {
-            Shape::Rect { ref p0, ref p1 } => {
-                if (p1.x - p0.x).abs() < (p1.y - p0.y).abs() {
-                    return Dir::Vert;
-                }
-                Dir::Horiz
-            }
-            // Polygon and Path elements always horizontal, at least for now
-            Shape::Poly { .. } | Shape::Path { .. } => Dir::Horiz,
-        }
-    }
-    /// Shift coordinates by the (x,y) values specified in `pt`
-    pub fn shift(&mut self, pt: &Point) {
-        match *self {
-            Shape::Rect {
-                ref mut p0,
-                ref mut p1,
-            } => {
-                p0.x += pt.x;
-                p0.y += pt.y;
-                p1.x += pt.x;
-                p1.y += pt.y;
-            }
-            Shape::Poly { ref mut pts } => {
-                for p in pts.iter_mut() {
-                    p.x += pt.x;
-                    p.y += pt.y;
-                }
-            }
-            Shape::Path { ref mut pts, .. } => {
-                for p in pts.iter_mut() {
-                    p.x += pt.x;
-                    p.y += pt.y;
-                }
-            }
-        }
-    }
-    /// Boolean indication of whether we contain point `pt`
-    pub fn contains(&self, pt: &Point) -> bool {
-        match self {
-            Shape::Rect { ref p0, ref p1 } => {
-                p0.x.min(p1.x) <= pt.x
-                    && p0.x.max(p1.x) >= pt.x
-                    && p0.y.min(p1.y) <= pt.y
-                    && p0.y.max(p1.y) >= pt.y
-            }
-            Shape::Poly { .. } => false, // FIXME! todo!(),
-            Shape::Path { ref width, ref pts } => {
-                // Break into segments, and check for intersection with each
-                // Probably not the most efficient way to do this, but a start.
-                // Only "Manhattan paths", i.e. those with segments solely running vertically or horizontally, are supported.
-                // FIXME: even with this method, there are some small pieces at corners which we'll miss.
-                // Whether these are relevant in real life, tbd.
-                let width = isize::try_from(*width).unwrap(); // FIXME: probably store these signed, check them on creation
-                for k in 0..pts.len() - 1 {
-                    let rect = if pts[k].x == pts[k + 1].x {
-                        Shape::Rect {
-                            p0: Point::new(pts[k].x - width / 2, pts[k].y),
-                            p1: Point::new(pts[k].x + width / 2, pts[k + 1].y),
-                        }
-                    } else if pts[k].y == pts[k + 1].y {
-                        Shape::Rect {
-                            p0: Point::new(pts[k].x, pts[k].y - width / 2),
-                            p1: Point::new(pts[k + 1].x, pts[k].y + width / 2),
-                        }
-                    } else {
-                        unimplemented!("Unsupported Non-Manhattan Path")
-                    };
-                    if rect.contains(pt) {
-                        return true;
-                    }
-                }
-                false
-            }
-        }
-    }
-}
