+++ lef21/src/write.rs
@@ -5,9 +5,14 @@
 // Standard Lib Imports
 use std::io::Write;
 use std::ops::{AddAssign, SubAssign};
+use std::path::Path;
+
+// Layout21 Imports
+use layout21utils as utils;
+pub use utils::{enumstr, EnumStr, SerdeFile, SerializationFormat};
 
 // Local imports
-use super::*;
+use super::data::*;
 
 /// Write a [LefLibrary] to file `fname`.  
 /// Fields are written in the LEF-recommended order.  
@@ -128,7 +133,7 @@ impl<'wr> LefWriter<'wr> {
             self.write_symmetries(v)?;
         }
         // ROWPATTERN would be written here
-        // if site.row_pattern.is_some() { } 
+        // if site.row_pattern.is_some() { }
         self.write_line(format_args_f!("{Size} {site.size.0} {By} {site.size.1} ;"))?;
         self.indent -= 1;
         self.write_line(format_args_f!("{End} {site.name} ; "))?;
